requirements.txt
langgraph
langsmith
langchain[openai]
langgraph-supervisor
sendgrid
hubspot-api-client